Updated the engine plugin implementation.
Now it does not use background threads, because QProcess is not thread safe.

I also made it much more functional and reliable, and it gives more information
in the log window.  Now you can see the messages it both sends and receives
with the engine.  It also correctly recovers from crashes.